# php/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-BookingHealthcare API - Swagger",
        "description": "E-BookingHealthcare API with Laravel",
        "contact": {
            "name": "Nguyen Chinh",
            "url": "https://github.com/ngchinhdev/",
            "email": "chinhnguyennn24@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/news": {
            "get": {
                "tags": [
                    "News Routes"
                ],
                "summary": "Get all News",
                "operationId": "a58f0ec26ff904bf768f333b0e274157",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/news/{id}": {
            "get": {
                "tags": [
                    "News Routes"
                ],
                "summary": "Get one News",
                "operationId": "d18c006333797f59c6e8c5dd2ec35d91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/news/add": {
            "post": {
                "tags": [
                    "News Routes"
                ],
                "summary": "Add a News",
                "operationId": "80644a98c634b91170845595e9ada5f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "specialtyID": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Title of the new News"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "Image of the new News"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "content of the new News"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "author of the new News"
                                    },
                                    "isHidden": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/news/update/{id}": {
            "put": {
                "tags": [
                    "News Routes"
                ],
                "summary": "Update News",
                "operationId": "8eb03544470fd03b6c2978d42733b0c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "object id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "specialtyID": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Title of the new News"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "Image of the new News"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "content of the new News"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "author of the new News"
                                    },
                                    "isHidden": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/news/delete/{id}": {
            "delete": {
                "tags": [
                    "News Routes"
                ],
                "summary": "Update News",
                "operationId": "deefd67e9cc9c49be602edfe36f4a2ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "object id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
...
    },
...
}